requirements.txt
requests
pytest
pact-python